+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Legend of version history:
 History:
 --------
 ### v1.1.6 -> v1.2.0 (2011/4/14)
+ - [*] 借鉴 jquery 1.6,支持 w3c attribute, attr 方法对 checked 等属性返回 "checked"
  - [!] DOM.insertBefore/insertAfter/append/prepend 没有返回值
  - [*] Node 与 NodeList 合一
  - [!] DOM 参数中不可以传入 KISSY Node 或 KISSY NodeList 类型